+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/WorldObserver.cpp
@@ -251,7 +251,12 @@ static carla::geom::Vector3D FWorldObserver_GetAngularVelocity(const AActor &Act
       RootComponent != nullptr ?
           RootComponent->GetPhysicsAngularVelocityInDegrees() :
           FVector{0.0f, 0.0f, 0.0f};
-  return {AngularVelocity.X, AngularVelocity.Y, AngularVelocity.Z};
+  return
+  {
+    static_cast<float>(AngularVelocity.X),
+    static_cast<float>(AngularVelocity.Y),
+    static_cast<float>(AngularVelocity.Z)
+  };
 }
 
 static carla::geom::Vector3D FWorldObserver_GetAcceleration(
@@ -262,7 +267,12 @@ static carla::geom::Vector3D FWorldObserver_GetAcceleration(
   FVector &PreviousVelocity = View.GetActorInfo()->Velocity;
   const FVector Acceleration = (Velocity - PreviousVelocity) / DeltaSeconds;
   PreviousVelocity = Velocity;
-  return {Acceleration.X, Acceleration.Y, Acceleration.Z};
+  return
+  {
+    static_cast<float>(Acceleration.X),
+    static_cast<float>(Acceleration.Y),
+    static_cast<float>(Acceleration.Z)
+  };
 }
 
 static carla::Buffer FWorldObserver_Serialize(
@@ -329,9 +339,11 @@ static carla::Buffer FWorldObserver_Serialize(
       const FActorData* ActorData = View->GetActorData();
       Velocity = TO_METERS * ActorData->Velocity;
       AngularVelocity = carla::geom::Vector3D
-                        {ActorData->AngularVelocity.X,
-                         ActorData->AngularVelocity.Y,
-                         ActorData->AngularVelocity.Z};
+      {
+        static_cast<float>(ActorData->AngularVelocity.X),
+        static_cast<float>(ActorData->AngularVelocity.Y),
+        static_cast<float>(ActorData->AngularVelocity.Z)
+      };
       Acceleration = FWorldObserver_GetAcceleration(*View, Velocity, DeltaSeconds);
       State = FWorldObserver_GetDormantActorState(*View, Registry);
     }
